refactor: 상품 더보기 버튼 관심사 분리
