+++ fe/src/components/ProductListItem.tsx
@@ -9,6 +9,14 @@ import { useMemberValue } from "store";
 import { styled } from "styled-components";
 import ProductMoreButton from "./ProductMoreButton";
 import { TextBold, TextDefault, TextWeak } from "@styles/common";
+import { useState } from "react";
+import {
+  useDeleteProductQuery,
+  useMutateProductStatus,
+  useProductStatusesQuery,
+} from "@api/product/queries";
+import Alert from "./common/Alert/Alert";
+import { MenuItemInfo } from "./common/Menu/type";
 
 type ProductListItemProps = {
   productItem: ProductItem;
@@ -21,65 +29,103 @@ export default function ProductListItem({
 }: ProductListItemProps) {
   const navigate = useNavigate();
   const { id: memberId } = useMemberValue();
+  const [isDeleteAlertOpen, setIsDeleteAlertOpen] = useState(false);
+
+  const { onDeleteProduct } = useDeleteProductQuery({
+    productId: productItem.productId,
+    invalidateQueryKey,
+  });
+  const { data: productStatuses, isSuccess } = useProductStatusesQuery();
+  const { mutate: mutateProductStatus } = useMutateProductStatus({
+    invalidateQueryKey,
+  });
 
   const onProductClick = () => {
-    navigate(`${ROUTE_PATH.detail}/${productItem.productId}`);
+    navigate(ROUTE_PATH.detail + `/${productItem.productId}`);
   };
+
+  const getProductStatusList = (): MenuItemInfo[] => {
+    return isSuccess
+      ? productStatuses
+          .filter((status) => status.id !== productItem.statusId)
+          .map((productStatus) => ({
+            name: `${productStatus.type} 상태로 전환`,
+            onClick: () => {
+              mutateProductStatus({
+                productId: productItem.productId,
+                statusId: productStatus.id,
+              });
+            },
+          }))
+      : [];
+  };
+
+  const openDeleteAlert = () => setIsDeleteAlertOpen(true);
+  const closeDeleteAlert = () => setIsDeleteAlertOpen(false);
 
   const isSeller = productItem.sellerId === memberId;
   const productPrice = productItem.price?.toLocaleString("ko-KR");
 
   return (
-    <StyledProductListItem onClick={onProductClick}>
-      <Product>
-        <ProductImage src={productItem.thumbnailUrl} />
-        <ProductInfo>
-          <div className="info-main">
-            <div className="info-top">
-              <TextDefault>{productItem.title}</TextDefault>
-              {isSeller && (
-                <ProductMoreButton
-                  productId={productItem.productId}
-                  currentStatusId={productItem.statusId}
-                  invalidateQueryKey={invalidateQueryKey}
-                />
+    <>
+      <StyledProductListItem onClick={onProductClick}>
+        <Product>
+          <ProductImage src={productItem.thumbnailUrl} />
+          <ProductInfo>
+            <div className="info-main">
+              <div className="info-top">
+                <TextDefault>{productItem.title}</TextDefault>
+                {isSeller && (
+                  <ProductMoreButton
+                    productId={productItem.productId}
+                    statusListItems={getProductStatusList()}
+                    openDeleteAlert={openDeleteAlert}
+                  />
+                )}
+              </div>
+              <div className="info-middle">
+                <TextWeak>{productItem.addressName}</TextWeak>
+                <TextWeak>・</TextWeak>
+                <TextWeak>
+                  {convertPastTimestamp(productItem.createdTime)}
+                </TextWeak>
+              </div>
+              <div className="info-bottom">
+                <ProductStatus id={productItem.statusId} />
+                <TextBold>
+                  {productPrice ? `${productPrice} 원` : "가격미정"}
+                </TextBold>
+              </div>
+            </div>
+            <div className="info-history">
+              {!!productItem.stats.chatCount && (
+                <>
+                  <History>
+                    <MessageIcon />
+                    <TextWeak>{productItem.stats.chatCount}</TextWeak>
+                  </History>
+                </>
+              )}
+              {!!productItem.stats.wishCount && (
+                <>
+                  <History>
+                    <HeartIcon />
+                    <TextWeak>{productItem.stats.wishCount}</TextWeak>
+                  </History>
+                </>
               )}
             </div>
-            <div className="info-middle">
-              <TextWeak>{productItem.addressName}</TextWeak>
-              <TextWeak>・</TextWeak>
-              <TextWeak>
-                {convertPastTimestamp(productItem.createdTime)}
-              </TextWeak>
-            </div>
-            <div className="info-bottom">
-              <ProductStatus id={productItem.statusId} />
-              <TextBold>
-                {productPrice ? `${productPrice} 원` : "가격미정"}
-              </TextBold>
-            </div>
-          </div>
-          <div className="info-history">
-            {!!productItem.stats.chatCount && (
-              <>
-                <History>
-                  <MessageIcon />
-                  <TextWeak>{productItem.stats.chatCount}</TextWeak>
-                </History>
-              </>
-            )}
-            {!!productItem.stats.wishCount && (
-              <>
-                <History>
-                  <HeartIcon />
-                  <TextWeak>{productItem.stats.wishCount}</TextWeak>
-                </History>
-              </>
-            )}
-          </div>
-        </ProductInfo>
-      </Product>
-    </StyledProductListItem>
+          </ProductInfo>
+        </Product>
+      </StyledProductListItem>
+      {isDeleteAlertOpen && (
+        <Alert
+          message={"상품을 삭제하시겠어요?"}
+          onDeleteClick={onDeleteProduct}
+          closeAlertHandler={closeDeleteAlert}
+        />
+      )}
+    </>
   );
 }
